grpc-js: Refactor code to eliminate runtime dependency cycles
