+++ packages/grpc-js/package.json
@@ -29,6 +29,7 @@
     "gulp": "^4.0.2",
     "gulp-mocha": "^6.0.0",
     "lodash": "^4.17.4",
+    "madge": "^5.0.1",
     "mocha-jenkins-reporter": "^0.4.1",
     "ncp": "^2.0.0",
     "pify": "^4.0.1",
@@ -53,7 +54,7 @@
     "check": "gts check src/**/*.ts",
     "fix": "gts fix src/*.ts",
     "pretest": "npm run compile",
-    "posttest": "npm run check"
+    "posttest": "npm run check && madge -c ./build/src"
   },
   "dependencies": {
     "@types/node": ">=12.12.47"
