+++ packages/grpc-js/src/load-balancer-pick-first.ts
@@ -18,10 +18,11 @@
 import {
   LoadBalancer,
   ChannelControlHelper,
-  registerLoadBalancerType,
-  LoadBalancingConfig
+  LoadBalancingConfig,
+  registerDefaultLoadBalancerType, 
+  registerLoadBalancerType
 } from './load-balancer';
-import { ConnectivityState } from './channel';
+import { ConnectivityState } from "./connectivity-state";
 import {
   QueuePicker,
   Picker,
@@ -33,9 +34,11 @@ import {
 import {
   Subchannel,
   ConnectivityStateListener,
-  SubchannelAddress,
-  subchannelAddressToString,
 } from './subchannel';
+import {
+  SubchannelAddress,
+  subchannelAddressToString
+} from "./subchannel-address";
 import * as logging from './logging';
 import { LogVerbosity } from './constants';
 
@@ -458,4 +461,5 @@ export class PickFirstLoadBalancer implements LoadBalancer {
 
 export function setup(): void {
   registerLoadBalancerType(TYPE_NAME, PickFirstLoadBalancer, PickFirstLoadBalancingConfig);
+  registerDefaultLoadBalancerType(TYPE_NAME);
 }
